chore: seperate function to apply filter for warehouse in case of QB
